feat: move auth store to wallet package
